invoice-web-app/requirements.txt
streamlit
requests
gspread
google-auth
google-api-python-client
pandas